school_api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
pydantic==2.11.4
pydantic-settings==2.9.1
python-jose==3.4.0
passlib==1.7.4
bcrypt==4.3.0
python-multipart==0.0.20
email-validator==2.2.0
python-dotenv==1.1.0
alembic==1.15.2 